% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{clustering_scores}
\alias{clustering_scores}
\title{Clustering Performance with Known Truth}
\usage{
clustering_scores(y, pred)
}
\arguments{
\item{y}{True clustering labels.}

\item{pred}{Cluster labels from model.}
}
\value{
Returns a list of performance scores.
}
\description{
Clustering Performance with Known Truth
}
