% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{importance_plot}
\alias{importance_plot}
\title{Plot the importance from an RZiMM object}
\usage{
importance_plot(rzimm_mod, cutoff = NULL, top = NULL)
}
\arguments{
\item{rzimm_mod}{An RZiMM object.}

\item{cutoff}{Cutoff to highlight important genes.}

\item{top}{Alternatively, one can choose to highlight top x genes.}
}
\value{
Returns a ggplot object.
}
\description{
Plot the importance from an RZiMM object
}
\seealso{
\code{\link{RZiMM-class}}\cr
}
