% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-RZiMM.R
\name{n_clust}
\alias{n_clust}
\title{Calculate how many different values in a vector (within a cutoff)}
\usage{
n_clust(v, cutoff)
}
\arguments{
\item{v}{A vector}

\item{cutoff}{The cutoff}
}
\value{
Returns a number object.
}
\description{
Calculate how many different values in a vector (within a cutoff)
}
