% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{optim_ind}
\alias{optim_ind}
\title{Match the cluster number of two clustering results, return the 
 matching numbers and the matching accuracy.}
\usage{
optim_ind(ind_pred, ind_true)
}
\arguments{
\item{ind_pred}{Clustering indices from one way.}

\item{ind_true}{Clustering indices from another way.}
}
\value{
Returns a list.
}
\description{
Match the cluster number of two clustering results, return the 
 matching numbers and the matching accuracy.
}
