% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{perf_stat}
\alias{perf_stat}
\title{Model performance in a simulated scenario}
\usage{
perf_stat(ind_est, imp_score, cut_off, true_ind, true_imp)
}
\arguments{
\item{ind_est}{Clustering result from an algorithm.}

\item{imp_score}{Importance scores from an algorithm.}

\item{cut_off}{Cutoff for importance score.}

\item{true_ind}{True clustering indices.}

\item{true_imp}{True importance (0 for unimportant; 1 for important).}
}
\value{
Returns a list of performance scores.
}
\description{
Model performance in a simulated scenario
}
