% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-2-generics.R
\name{rzi}
\alias{rzi}
\alias{rzimm_importance}
\alias{rzimm_importance,RZiMM-method}
\alias{rzimm_cluster}
\alias{rzimm_cluster,RZiMM-method}
\alias{rzimm_info}
\alias{rzimm_info,RZiMM-method}
\alias{rzimm_param}
\alias{rzimm_param,RZiMM-method}
\alias{rzimm_bic}
\alias{rzimm_bic,RZiMM-method}
\alias{rzimm_nonzero_prob}
\alias{rzimm_nonzero_prob,RZiMM-method}
\alias{rzimm_m}
\alias{rzimm_m,RZiMM-method}
\title{RZiMM Functions}
\usage{
rzimm_importance(object)

\S4method{rzimm_importance}{RZiMM}(object)

rzimm_cluster(object)

\S4method{rzimm_cluster}{RZiMM}(object)

rzimm_info(object)

\S4method{rzimm_info}{RZiMM}(object)

rzimm_param(object)

\S4method{rzimm_param}{RZiMM}(object)

rzimm_bic(object)

\S4method{rzimm_bic}{RZiMM}(object)

rzimm_nonzero_prob(object)

\S4method{rzimm_nonzero_prob}{RZiMM}(object)

rzimm_m(object)

\S4method{rzimm_m}{RZiMM}(object)
}
\arguments{
\item{object}{An \code{RZiMM} object.}
}
\description{
A collection of functions for an RZiMM object
}
\seealso{
\code{\link{RZiMM-class}}\cr
\code{\link{RZiMM}}\cr
}
