% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{RZiMM-class}
\alias{RZiMM-class}
\title{An S4 class containing predictors (x), response (y) and sample weights (w)}
\description{
An S4 class containing predictors (x), response (y) and sample weights (w)
}
\section{Slots}{

\describe{
\item{\code{cluster}}{A numeric vector of cluster number}

\item{\code{param}}{A list of parameters in the RZiMM model}

\item{\code{importance}}{A numeric vector for the importance of each gene}

\item{\code{info}}{A list of items, including log-likelihood, bic, and etc}
}}

\seealso{
\code{\link{RZiMM}}\cr
}
