% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-1-c-RZiMM-Count.R
\name{RZiMM_Count}
\alias{RZiMM_Count}
\title{A Regularized Zero-inflated Mixture Model for scRNA Data
  - for count data
  - Poisson or Negative Binomial}
\usage{
RZiMM_Count(
  x,
  mix_ind = NULL,
  n_group = 4,
  dist = c("zinb", "zip")[1],
  g_ind_init = NULL,
  n_iter = 100,
  n_iter_m = 10,
  lambda1 = 100,
  lambda2 = 0,
  epsilon0 = 10^-6,
  epsilon = 10^-6,
  print_e = TRUE,
  epsilon_p = 10^-3,
  ...
)
}
\arguments{
\item{x}{A n by p matrix, n cells x p genes.}

\item{mix_ind}{Sample index, for adjusting within-sample correlation; if set to be NULL,
the within-sample correlation will not be adjusted.}

\item{n_group}{Number of groups for clustering.}

\item{dist}{"zinb" for zero-inflated Negative Binomial or "zip" for zero-inflated Poisson.}

\item{g_ind_init}{Initial cluster index.}

\item{n_iter}{Maximum number of total iteration.}

\item{n_iter_m}{Number of iteration for the majorization step.}

\item{lambda1}{Weight of regularization for group effects.}

\item{lambda2}{Weight of regularization for sample effects.}

\item{epsilon0}{A small number to addess matrix singularity.}

\item{epsilon}{Iteration stops when the change from last iteration is smaller than this value.}

\item{print_e}{Whether to print the changes compared to the last iteration.}

\item{epsilon_p}{A small number for calculating the number of parameters.}

\item{...}{Other parameters passed to \code{zinb_regulated()} or \code{zip_regulated()}.}
}
\value{
Returns an \code{RZiMM-class} object.
}
\description{
A Regularized Zero-inflated Mixture Model for scRNA Data
  - for count data
  - Poisson or Negative Binomial
}
\seealso{
\code{\link{RZiMM-class}}\cr
}
