% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-methods.R
\name{show}
\alias{show}
\alias{show,RZiMM-method}
\title{Method show for the package}
\usage{
\S4method{show}{RZiMM}(object)
}
\arguments{
\item{object}{A \code{RZiMM} object.}
}
\description{
The method show for \code{RZiMM} object.
}
\seealso{
\code{\link{RZiMM-class}}\cr
}
