% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-3-data.R
\docType{data}
\name{sim_x}
\alias{sim_x}
\title{A simulated dataset}
\format{
A list
\describe{
  \item{x}{a data matrix: 1000 cells x 1000 genes, normalized, many zeros}
  \item{cluster}{cluster index}
  \item{sample_index}{sample index}
}
}
\usage{
sim_x
}
\description{
A list containing:
  a data matrix: 1000 cells x 1000 genes; 
  a vector: cluster index;
  a vector: sample index
}
\keyword{datasets}
