% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-3-data.R
\docType{data}
\name{sim_y}
\alias{sim_y}
\title{A simulated dataset (count)}
\format{
A list
\describe{
  \item{x}{a data matrix: 1000 cells x 1000 genes, normalized, many zeros; 
    The first 100 genes are DE genes for the cluster}
  \item{cluster}{cluster index}
  \item{sample_index}{sample index (batch effect)}
  \item{imp_gene_list}{indicator for important genes}
}

A list of data
}
\usage{
data(sim_y)
}
\description{
A list containing:
  a data matrix: 1000 cells x 1000 genes; 
  a vector: cluster index;
  a vector: sample index
  a vector: indicator for important genes
}
\keyword{datasets}
