% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{step2_sig}
\alias{step2_sig}
\title{Anova test followed by a clustering algorithm, e.g.
  K-Means, H-Clust, etc}
\usage{
step2_sig(x, ind_pred)
}
\arguments{
\item{x}{Data matrix.}

\item{ind_pred}{Clustering result.}
}
\value{
Returns a vector of p-values.
}
\description{
Anova test followed by a clustering algorithm, e.g.
  K-Means, H-Clust, etc
}
