% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-utils.R
\name{perf_stat}
\alias{perf_stat}
\title{DE Gene Identification Performance with Known Truth}
\usage{
perf_stat(imp_score, cut_off, true_imp)
}
\arguments{
\item{imp_score}{Importance scores from an algorithm.}

\item{cut_off}{Cutoff for importance score.}

\item{true_imp}{True importance (0 for unimportant; 1 for important).}
}
\value{
Returns a list of performance scores.
}
\description{
DE Gene Identification Performance with Known Truth
}
