% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-methods.R
\name{show}
\alias{show}
\alias{show,ZRiMM-method}
\title{Method show for the package}
\usage{
\S4method{show}{ZRiMM}(object)
}
\arguments{
\item{object}{A \code{ZRiMM} object.}
}
\description{
The method show for \code{ZRiMM} object.
}
\seealso{
\code{\link{ZRiMM-class}}\cr
}
